requirements.txt
pillow==10.2.0
diffusers==0.26.0
accelerate==0.26.1
transformers==4.37.2
torch==2.2.0
ffmpeg-python==0.2.0